requirements.txt
# Core Framework
fastapi
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Cloud Services
google-cloud-documentai
google-cloud-aiplatform
google-cloud-storage
google-cloud-translate
google-cloud-secret-manager
# AI/ML Libraries
transformers
torch
sentence-transformers
numpy
pandas

# Vector Database
faiss-cpu
pymilvus

# Database
sqlalchemy
alembic
psycopg2-binary

# Utilities
python-dotenv
requests
pydantic
jinja2
aiofiles
pydantic-settings

# Development
pytest
black
# Add new official SDKs:
google-genai>=1.1.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client
google-cloud-discoveryengine
numpy